+++ src/sys/class/alo/db/abstractdb.php
@@ -4,6 +4,7 @@
 
    use Alo;
    use Alo\Cache\AbstractCache;
+   use Log;
    use PDOStatement;
 
    if(!defined('GEN_START')) {
@@ -92,6 +93,10 @@
          function __construct($cache) {
             if(!\Alo::$cache) {
                $this->cache = new $cache;
+               Log::warning('Alo::$cache was not defined when ' . get_class($this) . ' was instantiated and got assigned a ' . $cache);
+            } elseif(!is_a(Alo::$cache, $cache)) {
+               $this->cache = new $cache;
+               Log::warning('Alo::$cache wasn\'t an instance of ' . $cache . ' when ' . get_class($this) . ' was instantiates and was overwritten.');
             } else {
                $this->cache = &\Alo::$cache;
             }